Moved related method to AdminCommands. Formatting fixes.
